DBZ-8585 Move SparseVector into debezium-core
