+++ debezium-connector-postgres/src/main/java/io/debezium/connector/postgresql/PostgresSchemaFactory.java
@@ -10,7 +10,6 @@ import org.apache.kafka.connect.data.SchemaBuilder;
 
 import io.debezium.config.CommonConnectorConfig;
 import io.debezium.connector.postgresql.data.Ltree;
-import io.debezium.connector.postgresql.data.vector.SparseVector;
 import io.debezium.data.Envelope;
 import io.debezium.schema.SchemaFactory;
 import io.debezium.schema.SchemaNameAdjuster;
@@ -73,15 +72,5 @@ public class PostgresSchemaFactory extends SchemaFactory {
         return SchemaBuilder.string()
                 .name(Ltree.LOGICAL_NAME)
                 .version(Ltree.SCHEMA_VERSION);
-    }
-
-    public SchemaBuilder datatypeSparseVectorSchema() {
-        return SchemaBuilder.struct()
-                .name(SparseVector.LOGICAL_NAME)
-                .name(SparseVector.LOGICAL_NAME)
-                .version(SparseVector.SCHEMA_VERSION)
-                .doc("Sparse vector")
-                .field(SparseVector.DIMENSIONS_FIELD, Schema.INT16_SCHEMA)
-                .field(SparseVector.VECTOR_FIELD, SchemaBuilder.map(Schema.INT16_SCHEMA, Schema.FLOAT64_SCHEMA).build());
     }
 }
