+++ debezium-core/src/main/java/io/debezium/schema/SchemaFactory.java
@@ -25,6 +25,7 @@ import io.debezium.data.VariableScaleDecimal;
 import io.debezium.data.Xml;
 import io.debezium.data.vector.DoubleVector;
 import io.debezium.data.vector.FloatVector;
+import io.debezium.data.vector.SparseVector;
 import io.debezium.heartbeat.HeartbeatImpl;
 import io.debezium.pipeline.notification.Notification;
 import io.debezium.pipeline.txmetadata.TransactionStructMaker;
@@ -377,6 +378,16 @@ public class SchemaFactory {
         return SchemaBuilder.array(Schema.FLOAT32_SCHEMA)
                 .name(FloatVector.LOGICAL_NAME)
                 .version(FloatVector.SCHEMA_VERSION);
+    }
+
+    public SchemaBuilder dataTypeSparseVectorSchema() {
+        return SchemaBuilder.struct()
+                .name(SparseVector.LOGICAL_NAME)
+                .name(SparseVector.LOGICAL_NAME)
+                .version(SparseVector.SCHEMA_VERSION)
+                .doc("Sparse vector")
+                .field(SparseVector.DIMENSIONS_FIELD, Schema.INT16_SCHEMA)
+                .field(SparseVector.VECTOR_FIELD, SchemaBuilder.map(Schema.INT16_SCHEMA, Schema.FLOAT64_SCHEMA).build());
     }
 
     public Envelope.Builder datatypeEnvelopeSchema() {
